+++ src/utils/hooks/API/useFeaturedBanners.js
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
-import { API_BASE_URL } from "../constants";
+import { API_BASE_URL } from "../../constants";
 import { useLatestAPI } from "./useLatestAPI";
 
-export function useProducts() {
+export function useFeaturedBanners() {
   const { ref: apiRef, isLoading: isApiMetadataLoading } = useLatestAPI();
   const [featuredBanners, setFeaturedBanners] = useState(() => ({
     data: {},
@@ -22,8 +22,8 @@ export function useProducts() {
 
         const response = await fetch(
           `${API_BASE_URL}/documents/search?ref=${apiRef}&q=${encodeURIComponent(
-            '[[at(document.type, "product")]]'
-          )}&lang=en-us&pageSize=12`,
+            '[[at(document.type, "banner")]]'
+          )}&lang=en-us&pageSize=5`,
           {
             signal: controller.signal,
           }
