+++ src/utils/hooks/API/useFeaturedProducts.js
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { API_BASE_URL } from "../constants";
+import { API_BASE_URL } from "../../constants";
 import { useLatestAPI } from "./useLatestAPI";
 
 export function useFeaturedProducts() {
